api/requirements_api.txt
pandas
fastapi
uvicorn
python-jose
passlib
pydantic
bcrypt
mysql-connector-python
python-dotenv
joblib
mlflow
python-multipart